requirements.txt
APScheduler==3.11.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
cmake==3.31.6
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
dialogflow==2.0.0
dj-database-url==2.3.0
Django==5.1.6
django-appconf==1.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crontab==0.7.1
django-imagekit==5.0.0
django-jazzmin==3.0.1
django-redis==5.4.0
django-render-block==0.10
django-templated-email==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
fonttools==4.57.0
google-api-core==1.34.1
google-auth==2.40.2
google-cloud-core==2.4.3
google-cloud-dialogflow==2.41.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.48.2
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==5.3.2
oscrypto==1.3.0
packaging==24.2
pilkit==3.0
pillow==11.1.0
proto-plus==1.26.1
protobuf==3.20.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pypdf==5.4.0
pyphen==0.17.2
python-bidi==0.6.6
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==8.1
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
uritools==4.0.3
urllib3==2.3.0
weasyprint==65.0
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
